# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Restaurante",
        "description": "Documentación de la API RESTful para el sistema de restaurante",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Servidor local"
        }
    ],
    "paths": {
        "/api/productos": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Listar todos los productos",
                "operationId": "496dc08306524d77f18d124d1d6838aa",
                "responses": {
                    "200": {
                        "description": "Listado de productos"
                    }
                }
            },
            "post": {
                "tags": [
                    "Productos"
                ],
                "summary": "Crear un nuevo producto",
                "operationId": "48b59f126c849beb4d4de4ff83b7945f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "precio"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Camiseta"
                                    },
                                    "precio": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 29.99
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Producto creado"
                    }
                }
            }
        },
        "/api/productos/{id}": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Obtener un producto por ID",
                "operationId": "672fde814ef8716ee902835cc96ae595",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Producto encontrado"
                    },
                    "404": {
                        "description": "Producto no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Productos"
                ],
                "summary": "Actualizar un producto existente",
                "operationId": "20b20d8e59992eb88abad7dcef3db6d7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Camisa actualizada"
                                    },
                                    "precio": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 39.99
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Producto actualizado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Productos"
                ],
                "summary": "Eliminar un producto por ID",
                "operationId": "4c0c7680428675134afd4d793f934f67",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Producto eliminado"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Productos",
            "description": "Operaciones relacionadas con productos"
        }
    ]
}